Revert "Add JSON-defined input format"
This reverts commit 274c407504f24e0b31ad0213398434f6f01a183b.
