--- shesmu-server/src/test/resources/test_json.shesmuschema
@@ -1,32 +1,0 @@
-{
-  "timeFormat": "ISO8660_STRING",
-  "variables": {
-    "x": {
-      "gangs": [
-        {
-          "dropIfDefault": false,
-          "gang": "hello",
-          "order": 0
-        }
-      ],
-      "signable": true,
-      "type": "i"
-    },
-    "y": {
-      "gangs": [
-        {
-          "dropIfDefault": false,
-          "gang": "hello",
-          "order": 1
-        }
-      ],
-      "signable": true,
-      "type": "s"
-    },
-    "z": {
-      "gangs": [],
-      "signable": false,
-      "type": "as"
-    }
-  }
-}
